cleaned up database setup
